fix: solucion a falla al subir archivos por chunks, algunos archivos con extensiones especificas(.rar por ejemplo) tenian errore al obtener el mime_type retornnando un 422.
solucion: crear funcion que extraiga la extension de forma personalizada

+++ app/Http/Requests/Storage/InitUploadRequest.php
@@ -16,7 +16,7 @@ class InitUploadRequest extends FormRequest
         return [
             "file_name" => ["required", "string", "max:255"],
             "mime_type" => ["required", "string", "max:127"],
-            "total_size" => ["required", "integer", "min:1"],
+            "total_size" => ["required", "numeric", "min:1"],
             "folder_id" => ["nullable", "string"],
         ];
     }
